+++ onnxscript/optimizer/constant_folding_test.py
@@ -1,30 +1,31 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
-import parameterized
 import unittest
 
 import onnx
+import parameterized
 import pytest
 
 import onnxscript.optimizer as optimizer
-from onnxscript.optimizer import constant_folding, constant_folding_ir
 from onnxscript.ir import serde
+from onnxscript.optimizer import constant_folding, constant_folding_ir
 
 
 @parameterized.parameterized_class(("using_ir",), [(False,), (True,)])
 class FoldConstantsTest(unittest.TestCase):
-
     def _fold(self, model: onnx.ModelProto, onnx_shape_inference=False):
         if self.using_ir:
             ir_model = serde.deserialize_model(model)
-            constant_folding_ir.fold_constants(ir_model, onnx_shape_inference=onnx_shape_inference)
+            constant_folding_ir.fold_constants(
+                ir_model, onnx_shape_inference=onnx_shape_inference
+            )
             optimizer.remove_unused_nodes(ir_model)
             return serde.serialize_model(ir_model)
         else:
             constant_folding.fold_constants(model, onnx_shape_inference=onnx_shape_inference)
             optimizer.remove_unused_nodes(model)
             return model
-    
+
     def test_fold_add(self):
         model = onnx.parser.parse_model(
             """
@@ -349,71 +350,6 @@ func (float[1,3] x) => ( return_val) {
         self.assertEqual(len(optimized.graph.node[1].output), 3)
         self.assertEqual(optimized.graph.node[1].op_type, "Split")
         self.assertEqual(len([n for n in optimized.graph.node if n.op_type == "Squeeze"]), 3)
-
-    def test_static_split_to_sequence_with_uneven_split(self):
-        model = onnx.parser.parse_model(
-            """
-<
-   ir_version: 8,
-   opset_import: ["pkg.onnxscript.torch_lib" : 1, "" : 18, "pkg.onnxscript.torch_lib.common" : 1],
-   producer_name: "pytorch",
-   producer_version: "2.2.0"
->
-main_graph (float[3,5] l_tensor_x_) => (float[3,5] return_val)
-   < _val_2, float[3,5] l_tensor_x_, float[2,5] getitem, float[1,5] getitem_1>
-{
-   _val_1 = Constant <value: tensor = int64 {2}> ()
-   _val_2 = pkg.onnxscript.torch_lib.aten_split <dim: int = 0> (l_tensor_x_, _val_1)
-   _val_3 = Constant <value: tensor = int64 {0}> ()
-   getitem = pkg.onnxscript.torch_lib.aten_getitem (_val_2, _val_3)
-   _val_5 = Constant <value: tensor = int64 {1}> ()
-   getitem_1 = pkg.onnxscript.torch_lib.aten_getitem (_val_2, _val_5)
-   return_val = Concat <axis: int = 0> (getitem_1, getitem)
-}
-<
-  domain: "pkg.onnxscript.torch_lib",
-  opset_import: ["" : 18]
->
-aten_split (self, split_size) => (return_val)
-{
-   return_val = SplitToSequence <axis: int = @dim> (self, split_size)
-}
-<
-  domain: "pkg.onnxscript.torch_lib",
-  opset_import: ["" : 18]
->
-aten_getitem (self, i) => (return_val)
-{
-   return_val = SequenceAt (self, i)
-}
-<
-  domain: "pkg.onnxscript.torch_lib.common",
-  opset_import: ["" : 18]
->
-Rank (input) => (return_val)
-{
-   tmp = Shape (input)
-   return_val = Size (tmp)
-}
-<
-  domain: "pkg.onnxscript.torch_lib.common",
-  opset_import: ["" : 18]
->
-IsScalar (input) => (return_val)
-{
-   tmp = Shape (input)
-   tmp_0 = Size (tmp)
-   tmp_1 = Constant <value_int: int = 0> ()
-   return_val = Equal (tmp_0, tmp_1)
-}
-            """
-        )
-        optimized = optimizer.optimize(model, num_iterations=1, onnx_shape_inference=False)
-
-        print(onnx.printer.to_text(optimized))
-        self.assertEqual(len(optimized.graph.node), 2)
-        self.assertEqual(len(optimized.graph.node[0].output), 2)
-        self.assertEqual(optimized.graph.node[0].op_type, "Split")
 
     def test_split_to_sequence_and_concat_from_sequence_with_new_axis_0(
         self,
